Add relative conditional/unconditional branches and subroutine calls
Modified the instruction set to better support relative conditional/unconditional branches and subroutine calls. If the addressing mode is immediate, then a conditional branch is modified to support a relative signed displacement. Similarly, the unconditional jump is modified to be an unconditional relative jump, and unconditional jump to subroutine is modified to a branch to subroutine. All relative computations treat the KA field as a signed 20-bit displacement added to the progam counter, which is pointing to the next instruction.